pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "git-roaster"
version = "0.1.0"
description = "Review your Github repo or user profile in unique, funny and sarcastic way"
readme = "README.md"
requires-python = ">=3.12"

# Add all your project's dependencies here
dependencies = [
    "typer-slim~=0.9",
    "pydantic-settings~=2.1",
    "PyGithub~=2.3",
    "beartype~=0.18",
    "cachetools~=5.3"
]

[project.scripts]
roast = "main:app"
